Fang 10.00 12062024

+++ Models/ResultGrpProductOverviewModel.cs
@@ -5,7 +5,7 @@
 
         //Display Box
         public int SumEmp { get; set; } //View_EFFReport  => sum(CountQRCode)
-        public int SumFG { get; set; }
+        public decimal SumFG { get; set; }
         public decimal CapHr { get; set; } //View_EFFReport  => sum(FinidGood)/sum(CountQRCode)
 
 
@@ -31,7 +31,7 @@
         public decimal YieldTarget { get; set; } //View_EFFReport.PercentYield
         public decimal YieldActual { get; set; } //View_EFFReport.YieldDefect
         public decimal DiffYield { get; set; } //View_EFFReport => YieldDefect - PercentYield
-        public int TotalDefect { get; set; } //View_EFFReport => Defect
+        public decimal TotalDefect { get; set; } //View_EFFReport => Defect
 
 
     }
